[#22] 친구추가 네비게이션 설정
