+++ Shoak/Actors/AppleAdminActor/UseCases/AppleUseCase.swift
@@ -8,10 +8,8 @@
 import Foundation
 import SwiftUI
 import AuthenticationServices
-import Messages
-import MessageUI
 
-class AppleUseCase: NSObject, MFMessageComposeViewControllerDelegate  {
+class AppleUseCase{
     
     // Sign in with Apple
     func handleSignInSuccess(authorization: ASAuthorization) {
@@ -31,28 +29,5 @@ class AppleUseCase: NSObject, MFMessageComposeViewControllerDelegate  {
     
     func handleSignInFailure(error: Error) {
         print(error)
-    }
-    
-    // 초대 행위를 전달하기 (iMessage)
-    func createMessageComposeViewController() -> MFMessageComposeViewController? {
-        let messageComposeVC = MFMessageComposeViewController()
-        let templateLayout = MSMessageTemplateLayout()
-        templateLayout.image = UIImage(systemName: "circle")
-        templateLayout.caption = "캡션"
-        let layout = MSMessageLiveLayout(alternateLayout: templateLayout)
-        
-        let message = MSMessage()
-        message.layout = layout
-        message.summaryText = "shoak 친구 초대 메세지"
-        
-        messageComposeVC.message = message
-        messageComposeVC.body = "친구 초대 메세지입니다."
-        messageComposeVC.messageComposeDelegate = self
-        
-        return messageComposeVC
-    }
-    
-    func messageComposeViewController(_ controller: MFMessageComposeViewController, didFinishWith result: MessageComposeResult) {
-        controller.dismiss(animated: true, completion: nil)
     }
 }
