--- Shoak/Actors/InvitationActor/UseCases/SampleUseCase2.swift
@@ -1,1 +1,0 @@
-
